@@ -4,75 +4,55 @@
   "description": "",
   "main": "./dist/index.js",
   "scripts": {
-    "watch-ts": "tsc -w",
+    "watch": "ts-node-dev --watch 'src' --respawn --transpile-only src/index.ts",
     "start": "node ./dist/index.js",
-    "server": "nodemon ./dist/index.js",
+    "server": "node ./dist/index.js",
     "client": "npm start --prefix ../client",
     "dev": "concurrently  \"npm run server\" \"npm run client \""
   },
   "author": "",
-  "license": "ISC",
+  "private": true,
+  "license": "UNLICENSED",
   "dependencies": {
-    "bcrypt": "^3.0.8",
-    "body-parser": "^1.19.0",
-    "concurrently": "^4.1.2",
-    "crypto": "^1.0.1",
+    "bcrypt": "^5.0.0",
+    "concurrently": "^5.3.0",
     "express": "^4.17.1",
     "fbgraph": "^1.4.4",
-    "googleapis": "^47.0.0",
+    "googleapis": "^67.0.0",
     "gridfs-stream": "^1.1.1",
     "jsonwebtoken": "^8.5.1",
-    "jwt-decode": "^2.2.0",
-    "method-override": "^3.0.0",
-    "mongoose": "^5.9.9",
+    "mongoose": "^5.11.14",
     "multer": "^1.4.2",
     "multer-gridfs-storage": "^4.2.0",
-    "nodemailer": "^6.4.10",
-    "passport": "^0.4.0",
+    "nodemailer": "^6.4.17",
+    "passport": "^0.4.1",
     "passport-local": "^1.0.0",
-    "path": "^0.12.7",
-    "socket.io": "^2.3.0",
-    "socket.io-client": "^2.3.0",
-    "validator": "^10.11.0"
+    "socket.io": "^3.1.0",
+    "socket.io-client": "^3.1.0",
+    "validator": "^13.5.2"
   },
   "devDependencies": {
-    "@types/async": "^3.2.0",
-    "@types/bcrypt-nodejs": "^0.0.30",
-    "@types/bluebird": "^3.5.27",
-    "@types/body-parser": "^1.17.1",
-    "@types/chai": "^4.2.11",
-    "@types/compression": "^1.0.1",
-    "@types/concurrently": "^4.1.0",
-    "@types/connect-mongo": "^0.0.43",
-    "@types/dotenv": "^6.1.1",
-    "@types/errorhandler": "^0.0.32",
-    "@types/eslint": "^6.8.0",
-    "@types/express": "^4.17.6",
-    "@types/express-flash": "0.0.1",
-    "@types/express-session": "^1.15.14",
-    "@types/jest": "^24.0.18",
-    "@types/jquery": "^3.3.34",
-    "@types/lodash": "^4.14.141",
-    "@types/lusca": "^1.6.1",
-    "@types/mongoose": "^5.7.10",
-    "@types/node": "^12.12.35",
-    "@types/node-sass": "^4.11.0",
-    "@types/nodemailer": "^6.2.1",
-    "@types/passport": "^1.0.3",
-    "@types/passport-facebook": "^2.1.9",
+    "@types/bcrypt-nodejs": "^0.0.31",
+    "@types/concurrently": "^5.2.1",
+    "@types/express": "^4.17.11",
+    "@types/mongoose": "^5.10.3",
+    "@types/node": "^14.14.22",
+    "@types/nodemailer": "^6.4.0",
+    "@types/passport": "^1.0.5",
+    "@types/passport-facebook": "^2.1.10",
     "@types/passport-local": "^1.0.33",
-    "@types/pug": "^2.0.4",
-    "@types/request": "^2.48.3",
-    "@types/request-promise": "^4.1.46",
-    "@types/shelljs": "^0.8.7",
-    "@types/supertest": "^2.0.8",
-    "@types/winston": "^2.4.4",
-    "@typescript-eslint/eslint-plugin": "^2.28.0",
-    "@typescript-eslint/parser": "^2.28.0",
+    "@typescript-eslint/eslint-plugin": "^4.14.2",
+    "@typescript-eslint/parser": "^4.14.2",
+    "eslint": "^7.19.0",
+    "eslint-plugin-prettier": "^3.3.1",
     "morgan": "^1.10.0",
-    "nodemon": "^1.19.4",
-    "ts-node": "^8.8.2",
-    "ts-node-dev": "^1.0.0-pre.44",
-    "typescript": "^3.6.3"
+    "prettier": "^2.2.1",
+    "ts-node-dev": "^1.1.1",
+    "typescript": "^4.1.3"
+  },
+  "prettier" {
+    "singleQuote": true,
+    "trailingComma": "all",
+    "arrowParens": "always"
   }
 }
